Correct counts by service status
Services can have multiple health checks associated with them.  The
worst status of any of the checks should be the status of the service as
a whole.  If any health check reports `critical`, increment the
`critical` count of the service.  Otherwise, if any health check reports
`warning`, increment the `warning` count of the service.  Otherwise,
increment the `passing` count of the service.